Move UpdateHandler to separate file
