+++ src/config/netwatch.ini
@@ -1,6 +1,6 @@
 # Start Main Menu Configurations
 [general_config]
-__version__ = 3.1.0
+__version__ = 3.1.4
 prompt = netwatch ~#
 config_file_path = ./config.ini
 logger_config_path = ./logger.yaml
